requirements.txt
fastapi
python-dotenv
pymongo